backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.3",
    "pillow>=10.4.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.1.1",
    "python-json-logger>=2.0.7",
    "python-multipart>=0.0.12",
    "supabase>=2.17.0",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "faker>=37.4.2",
    "httpx>=0.28.1",
    "mypy>=1.17.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-env>=1.1.5",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.4",
]
